+++ front/src/features/roasters/components/organisms/RoasterCard.tsx
@@ -5,7 +5,7 @@ import { Link as ReactLink } from 'react-router-dom'
 import { Card, CardContainer } from '@/components/Elements/Card'
 import { Link } from '@/components/Elements/Link'
 import { useAuth } from '@/features/auth'
-import { FollowButton, getRoasterRelationship } from '@/features/roaster_relationships'
+import { FollowButton, getRoasterRelationship } from '@/features/roasterRelationships'
 import { LinkToRoasterFollower } from '@/features/roasters/components/molecules/LinkToRoasterFollower'
 import { RoasterImage } from '@/features/roasters/components/molecules/RoasterImage'
 import { useCurrentRoaster } from '@/features/roasters/hooks/useCurrentRoaster'
